fix: cast logits to float32 before cross_entropy in pure bf16 mode
PyTorch cross_entropy does not support bfloat16 inputs. In pure bf16
training mode, model outputs are bfloat16 causing RuntimeError. Cast
pred.float() at all 4 cross_entropy call sites: main CE, deep
supervision CE, edge loss CE, and class weights.

+++ losses/edge_loss.py
@@ -44,6 +44,6 @@ class EdgeLoss(nn.Module):
         """
         edge_mask = self.get_edge_mask(target)  # [B, 1, D, H, W]
         edge_weight_map = edge_mask.squeeze(1)  # [B, D, H, W]
-        loss = F.cross_entropy(pred, target, reduction='none')  # [B, D, H, W]
+        loss = F.cross_entropy(pred.float(), target, reduction='none')  # [B, D, H, W]
         edge_bonus = (loss * edge_weight_map).mean()
         return edge_bonus
